old-tests/submission/W3C/domhtml/HTMLLinkElement06.html

<!DOCTYPE html>
<html><head>
<meta charset="utf-8">
<title>link</title>
<link href="/support/html/sample.html" hreflang="en" media="screen" rel="Glossary" type="text/html">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<link href="http://www.w3.org/TR/html5/the-link-element.html#dom-link-rel" rel="help">
<link rel="author" title="W3C">
</head>
<body>
<p>The rel attribute specifies the forward link type.     

    Retrieve the rel attribute and examine its value.</p>

<script>
var domtest = async_test("The rel attribute specifies the forward link type.");

on_event(window, "load", domtest.step_func(function (dom_event) {
    var nodeList;
      var testNode;
      var vrel;
      var doc;
      doc = window.document;
      nodeList = doc.getElementsByTagName("link");
      assert_equals(nodeList.length,3);
testNode = nodeList.item(0);
      vrel = testNode.rel;

      assert_equals(vrel,"Glossary");
       
    domtest.done();
}));

</script>
<div id="log"></div>
</body></html>
